refactor(openai): fold STT provider into speech provider
