+++ Examples/LangTools_Example/Modules/Audio/STTService.swift
@@ -11,7 +11,7 @@ import SwiftUI
 import Combine
 import AVFoundation
 import Speech
-import class OpenAI.OpenAISTTProvider
+import class OpenAI.OpenAISpeechRecognitionProvider
 
 /// STT processing status for UI feedback
 public enum STTStatus: Equatable {
@@ -133,7 +133,7 @@ public class STTService: ObservableObject {
         }
 
         // If switching to OpenAI, refresh API key from keychain
-        if type == .openAIWhisper, let openAIProvider = providers[.openAIWhisper] as? OpenAISTTProvider {
+        if type == .openAIWhisper, let openAIProvider = providers[.openAIWhisper] as? OpenAISpeechRecognitionProvider {
             print("[STTService] Refreshing OpenAI API key...")
             openAIProvider.refreshApiKey()
         }
