+++ Sources/OpenAI/OpenAISpeechRecognitionProvider.swift
@@ -2,8 +2,14 @@ import Foundation
 import LangTools
 
 /// Reusable OpenAI speech-to-text provider adapter.
+///
+/// This provider intentionally does not perform audio transcoding. Callers must
+/// pass audio bytes in a format supported by OpenAI's transcription endpoint and
+/// either configure `defaultFileType` or call `transcribe(audioData:fileType:...)`.
 @MainActor
-public final class OpenAISpeechRecognitionProvider: SpeechRecognitionProviding {
+open class OpenAISpeechRecognitionProvider: SpeechRecognitionProviding {
+    public typealias AudioInputNormalizer = @MainActor (Data) throws -> (audioData: Data, fileType: OpenAI.AudioTranscriptionRequest.FileType)
+
     public let providerID = LangToolsProviderID(rawValue: "openai.whisper")
     public let displayName = "OpenAI Whisper"
     public let capabilities = ProviderCapabilities(
@@ -19,7 +25,10 @@ public final class OpenAISpeechRecognitionProvider: SpeechRecognitionProviding {
     public private(set) var currentTranscript = ""
 
     private var openAI: OpenAI?
-    private var languageIdentifier: String?
+    private var defaultFileType: OpenAI.AudioTranscriptionRequest.FileType
+    private var languageIdentifierProvider: @MainActor () -> String?
+    private var apiKeyProvider: @MainActor () -> String?
+    private var audioInputNormalizer: AudioInputNormalizer?
 
     public var authorizationState: ProviderAuthorizationState {
         openAI == nil ? .unavailable(reason: "Missing OpenAI client") : .authorized
@@ -29,21 +38,79 @@ public final class OpenAISpeechRecognitionProvider: SpeechRecognitionProviding {
     public var isAvailable: Bool { openAI != nil }
     public var isListening: Bool { false }
 
-    public init(openAI: OpenAI? = nil, languageIdentifier: String? = nil) {
+    public init(
+        openAI: OpenAI? = nil,
+        defaultFileType: OpenAI.AudioTranscriptionRequest.FileType = .wav,
+        languageIdentifierProvider: @escaping @MainActor () -> String? = { nil },
+        audioInputNormalizer: AudioInputNormalizer? = nil
+    ) {
         self.openAI = openAI
-        self.languageIdentifier = languageIdentifier
+        self.defaultFileType = defaultFileType
+        self.languageIdentifierProvider = languageIdentifierProvider
+        self.apiKeyProvider = { nil }
+        self.audioInputNormalizer = audioInputNormalizer
+    }
+
+    public convenience init(
+        openAI: OpenAI? = nil,
+        languageIdentifier: String?
+    ) {
+        self.init(
+            openAI: openAI,
+            languageIdentifierProvider: { languageIdentifier }
+        )
+    }
+
+    public convenience init(
+        apiKeyProvider: @escaping @MainActor () -> String?,
+        defaultFileType: OpenAI.AudioTranscriptionRequest.FileType = .wav,
+        languageIdentifierProvider: @escaping @MainActor () -> String? = { nil },
+        audioInputNormalizer: AudioInputNormalizer? = nil
+    ) {
+        self.init(
+            openAI: apiKeyProvider().map { OpenAI(apiKey: $0) },
+            defaultFileType: defaultFileType,
+            languageIdentifierProvider: languageIdentifierProvider,
+            audioInputNormalizer: audioInputNormalizer
+        )
+        self.apiKeyProvider = apiKeyProvider
     }
 
     public func updateOpenAI(_ openAI: OpenAI?) {
+        apiKeyProvider = { nil }
         self.openAI = openAI
     }
 
-    public func configure(languageIdentifier: String) {
-        self.languageIdentifier = languageIdentifier == "auto" ? nil : languageIdentifier
+    public func updateApiKey(_ apiKey: String) {
+        apiKeyProvider = { apiKey }
+        openAI = OpenAI(apiKey: apiKey)
     }
 
-    public func requestAuthorization() async -> ProviderAuthorizationState { authorizationState }
-    public func refreshAuthorizationState() {}
+    public func refreshApiKey() {
+        // A nil dynamic key means "keep the currently injected client"; callers can
+        // explicitly clear credentials with updateOpenAI(nil).
+        guard let apiKey = apiKeyProvider() else { return }
+        openAI = OpenAI(apiKey: apiKey)
+    }
+
+    public func updateDefaultFileType(_ fileType: OpenAI.AudioTranscriptionRequest.FileType) {
+        defaultFileType = fileType
+    }
+
+    public func configure(languageIdentifier: String) {
+        let normalizedLanguageIdentifier = languageIdentifier == "auto" ? nil : languageIdentifier
+        languageIdentifierProvider = { normalizedLanguageIdentifier }
+    }
+
+    public func requestAuthorization() async -> ProviderAuthorizationState {
+        refreshApiKey()
+        return authorizationState
+    }
+
+    public func refreshAuthorizationState() {
+        refreshApiKey()
+    }
+
     public func prepareAssetsIfNeeded() {}
 
     public func startRecognition() throws {
@@ -61,7 +128,12 @@ public final class OpenAISpeechRecognitionProvider: SpeechRecognitionProviding {
     public func finalizeRecognition() {}
 
     public func transcribe(audioData: Data) async throws -> any LangToolsTranscriptionResponse {
-        try await transcribe(audioData: audioData, fileType: .wav, language: languageIdentifier)
+        refreshApiKey()
+        if let audioInputNormalizer {
+            let normalizedInput = try audioInputNormalizer(audioData)
+            return try await transcribe(audioData: normalizedInput.audioData, fileType: normalizedInput.fileType)
+        }
+        return try await transcribe(audioData: audioData, fileType: defaultFileType)
     }
 
     public func transcribe(
@@ -77,7 +149,7 @@ public final class OpenAISpeechRecognitionProvider: SpeechRecognitionProviding {
             file: audioData,
             fileType: fileType,
             prompt: prompt,
-            language: language ?? languageIdentifier,
+            language: language ?? languageIdentifierProvider(),
             responseFormat: responseFormat
         )
         let response = try await openAI.perform(request: request)
